Correct several issues regarding incorrect calculation of speed/acell

+++ src/vt/tr_types.h
@@ -452,11 +452,11 @@ typedef struct
     uint8_t frame_size;         // number of bit16's in Frames[] used by this animation
     uint16_t state_id;
 
-    float   speed;
-    float   accel;
+    int32_t   speed;
+    int32_t   accel;
 
-    float   speed_lateral;      // new in TR4 -->
-    float   accel_lateral;      // lateral speed and acceleration.
+    int32_t   speed_lateral;      // new in TR4 -->
+    int32_t   accel_lateral;      // lateral speed and acceleration.
 
     uint16_t frame_start;           // first frame in this animation
     uint16_t frame_end;             // last frame in this animation (numframes = (End - Start) + 1)
